// src/generated/api.ts
export interface BaseItemSpec {
    'code': string;
    'title': string;
}

export const Category = {
    Wearable: 'WEARABLE',
    Useable: 'USEABLE'
} as const;

export type Category = typeof Category[keyof typeof Category];


export interface CategorySpecDTO {
    'code': Category;
    'title': string;
}


export interface CreateItemRequestDTO {
    'name': string;
    'description'?: string;
    'priceCurrency'?: Currency;
    'askingPrice'?: number;
    'owner': ItemOwnerDTO;
    'category': Category;
    'slot'?: Slot;
    'weaponType'?: WeaponType;
    'element'?: Element;
    'level'?: number;
    'gender'?: Gender;
    'stats'?: Array<StatEntryDTO>;
}


export interface CreateItemResponseDTO {
    /**
     * Document id
     */
    '_id': string;
    'name': string;
    'description'?: string;
    'image'?: string;
    'priceCurrency'?: string;
    'askingPrice'?: number;
    'owner': ItemOwnerDTO;
    'category': string;
    'slot'?: string;
    'weaponType'?: string;
    'element'?: string;
    'level'?: number;
    'gender'?: string;
    'stats'?: Array<StatEntryDTO>;
}

export const Currency = {
    Silver: 'SILVER',
    Gold: 'GOLD',
    Coin: 'COIN'
} as const;

export type Currency = typeof Currency[keyof typeof Currency];


export interface CurrencySpecDTO {
    'code': Currency;
    'title': string;
}



export const Element = {
    Metal: 'METAL',
    Wood: 'WOOD',
    Water: 'WATER',
    Fire: 'FIRE',
    Earth: 'EARTH'
} as const;

export type Element = typeof Element[keyof typeof Element];


export interface ElementSpecDTO {
    'code': Element;
    'title': string;
}



export const GameServer = {
    S7Thaison: 'S7_THAISON'
} as const;

export type GameServer = typeof GameServer[keyof typeof GameServer];



export const Gender = {
    Male: 'MALE',
    Female: 'FEMALE'
} as const;

export type Gender = typeof Gender[keyof typeof Gender];


export interface GenderSpecDTO {
    'code': Gender;
    'title': string;
}


export interface GetItemSpecResponseDTO {
    'server': Array<BaseItemSpec>;
    'category': Array<CategorySpecDTO>;
    'slot': Array<SlotSpecDTO>;
    'weaponType': Array<WeaponTypeSpecDTO>;
    'element': Array<ElementSpecDTO>;
    'stat': Array<StatSpecDTO>;
    'currency': Array<CurrencySpecDTO>;
    'gender': Array<GenderSpecDTO>;
}
// ...
export interface ItemOwnerDTO {
    'server': GameServer;
    'playerName': string;
    'contactInfo'?: string;
}


export interface ListResponseDTO {
    'items': Array<string>;
    'hasNext': boolean;
    'skip': number;
    'limit': number;
}
export interface ResponseDTO {
    'statusCode': number;
    'message': string;
}

export const Slot = {
    Head: 'HEAD',
    Necklace: 'NECKLACE',
    Weapon: 'WEAPON',
    Armor: 'ARMOR',
    Gloves: 'GLOVES',
    Ring: 'RING',
    Amulet: 'AMULET',
    Belt: 'BELT',
    Boots: 'BOOTS'
} as const;

export type Slot = typeof Slot[keyof typeof Slot];


export interface SlotSpecDTO {
    'code': Slot;
    'title': string;
}



export const Stat = {
    AllResistance: 'ALL_RESISTANCE',
    MetalResistance: 'METAL_RESISTANCE',
    WoodResistance: 'WOOD_RESISTANCE',
    WaterResistance: 'WATER_RESISTANCE',
    FireResistance: 'FIRE_RESISTANCE',
    EarthResistance: 'EARTH_RESISTANCE',
    MetalExternalFlatAttack: 'METAL_EXTERNAL_FLAT_ATTACK',
    MetalExternalPercentAttack: 'METAL_EXTERNAL_PERCENT_ATTACK',
    MetalInternalFlatAttack: 'METAL_INTERNAL_FLAT_ATTACK',
    MetalInternalPercentAttack: 'METAL_INTERNAL_PERCENT_ATTACK',
    WoodExternalFlatAttack: 'WOOD_EXTERNAL_FLAT_ATTACK',
    WoodExternalPercentAttack: 'WOOD_EXTERNAL_PERCENT_ATTACK',
    WoodInternalFlatAttack: 'WOOD_INTERNAL_FLAT_ATTACK',
    WoodInternalPercentAttack: 'WOOD_INTERNAL_PERCENT_ATTACK',
    WaterExternalFlatAttack: 'WATER_EXTERNAL_FLAT_ATTACK',
    WaterExternalPercentAttack: 'WATER_EXTERNAL_PERCENT_ATTACK',
    WaterInternalFlatAttack: 'WATER_INTERNAL_FLAT_ATTACK',
    WaterInternalPercentAttack: 'WATER_INTERNAL_PERCENT_ATTACK',
    FireExternalFlatAttack: 'FIRE_EXTERNAL_FLAT_ATTACK',
    FireExternalPercentAttack: 'FIRE_EXTERNAL_PERCENT_ATTACK',
    FireInternalFlatAttack: 'FIRE_INTERNAL_FLAT_ATTACK',
    FireInternalPercentAttack: 'FIRE_INTERNAL_PERCENT_ATTACK',
    EarthExternalFlatAttack: 'EARTH_EXTERNAL_FLAT_ATTACK',
    EarthExternalPercentAttack: 'EARTH_EXTERNAL_PERCENT_ATTACK',
    EarthInternalFlatAttack: 'EARTH_INTERNAL_FLAT_ATTACK',
    EarthInternalPercentAttack: 'EARTH_INTERNAL_PERCENT_ATTACK',
    MaxHealthPointsFlat: 'MAX_HEALTH_POINTS_FLAT',
    MaxManaPointsFlat: 'MAX_MANA_POINTS_FLAT',
    MaxStaminaPointsFlat: 'MAX_STAMINA_POINTS_FLAT',
    CriticalChance: 'CRITICAL_CHANCE',
    CriticalDamage: 'CRITICAL_DAMAGE',
    CriticalResistance: 'CRITICAL_RESISTANCE',
    CriticalBlock: 'CRITICAL_BLOCK',
    HealthPointsRegeneration: 'HEALTH_POINTS_REGENERATION',
    ManaPointsRegeneration: 'MANA_POINTS_REGENERATION',
    Strength: 'STRENGTH',
    Vitality: 'VITALITY',
    Agility: 'AGILITY',
    Intellect: 'INTELLECT',
    Luck: 'LUCK',
    SlowReduction: 'SLOW_REDUCTION',
    StunReduction: 'STUN_REDUCTION',
    KnockbackReduction: 'KNOCKBACK_REDUCTION',
    PoisonReduction: 'POISON_REDUCTION',
    SlowBonus: 'SLOW_BONUS',
    StunBonus: 'STUN_BONUS',
    KnockbackBonus: 'KNOCKBACK_BONUS',
    PoisonBonus: 'POISON_BONUS',
    MoveSpeed: 'MOVE_SPEED',
    InternalAttackSpeed: 'INTERNAL_ATTACK_SPEED',
    ExternalAttackSpeed: 'EXTERNAL_ATTACK_SPEED'
} as const;

export type Stat = typeof Stat[keyof typeof Stat];


export interface StatEntryDTO {
    'code': Stat;
    'value': number;
}


export interface StatSpecDTO {
    'code': Stat;
    'title': string;
}



export const WeaponType = {
    Staff: 'STAFF',
    Mace: 'MACE',
    Blade: 'BLADE',
    DualBlade: 'DUAL_BLADE',
    Sword: 'SWORD',
    ThrowingBlade: 'THROWING_BLADE',
    ThrowingDart: 'THROWING_DART',
    Crossbow: 'CROSSBOW',
    Spear: 'SPEAR',
    Gauntlet: 'GAUNTLET'
} as const;

export type WeaponType = typeof WeaponType[keyof typeof WeaponType];


export interface WeaponTypeSpecDTO {
    'code': WeaponType;
    'title': string;
}
// ...
export interface InventoryApiInventoryControllerCreateItemV1Request {
    readonly createItemRequestDTO: CreateItemRequestDTO
}

/**
 * Request parameters for inventoryControllerDeleteItemV1 operation in InventoryApi.
 */
export interface InventoryApiInventoryControllerDeleteItemV1Request {
    readonly id: string
}

/**
 * Request parameters for inventoryControllerGetItemV1 operation in InventoryApi.
 */
export interface InventoryApiInventoryControllerGetItemV1Request {
    readonly id: string
}

/**
 * Request parameters for inventoryControllerGetItemsV1 operation in InventoryApi.
 */
export interface InventoryApiInventoryControllerGetItemsV1Request {
    /**
     * Array of { key, order }. Order in array &#x3D; sort priority (first &#x3D; primary). JSON string or array.
     */
    readonly sort?: string

    /**
     * Number of items to skip (offset)
     */
    readonly skip?: number

    /**
     * Page size
     */
    readonly limit?: number

    /**
     * JSON string of field filters. Ops per field: eq, gte, lte, gt, lt, in, match.
     */
    readonly query?: string
}

/**
 * Request parameters for inventoryControllerUploadItemImageV1 operation in InventoryApi.
 */
export interface InventoryApiInventoryControllerUploadItemImageV1Request {
    readonly id: string

    readonly image?: File
}

/**
 * InventoryApi - object-oriented interface
 */
export class InventoryApi extends BaseAPI {
    /**
     * 
     * @param {InventoryApiInventoryControllerCreateItemV1Request} requestParameters Request parameters.
     * @param {*} [options] Override http request option.
     * @throws {RequiredError}
     */
    public inventoryControllerCreateItemV1(requestParameters: InventoryApiInventoryControllerCreateItemV1Request, options?: RawAxiosRequestConfig) {
        return InventoryApiFp(this.configuration).inventoryControllerCreateItemV1(requestParameters.createItemRequestDTO, options).then((request) => request(this.axios, this.basePath));
    }

    /**
     * 
     * @param {InventoryApiInventoryControllerDeleteItemV1Request} requestParameters Request parameters.
     * @param {*} [options] Override http request option.
     * @throws {RequiredError}
     */
    public inventoryControllerDeleteItemV1(requestParameters: InventoryApiInventoryControllerDeleteItemV1Request, options?: RawAxiosRequestConfig) {
        return InventoryApiFp(this.configuration).inventoryControllerDeleteItemV1(requestParameters.id, options).then((request) => request(this.axios, this.basePath));
    }

    /**
     * 
     * @param {*} [options] Override http request option.
     * @throws {RequiredError}
     */
    public inventoryControllerGetItemSpecsV1(options?: RawAxiosRequestConfig) {
        return InventoryApiFp(this.configuration).inventoryControllerGetItemSpecsV1(options).then((request) => request(this.axios, this.basePath));
    }

    /**
     * 
     * @param {InventoryApiInventoryControllerGetItemV1Request} requestParameters Request parameters.
     * @param {*} [options] Override http request option.
     * @throws {RequiredError}
     */
    public inventoryControllerGetItemV1(requestParameters: InventoryApiInventoryControllerGetItemV1Request, options?: RawAxiosRequestConfig) {
        return InventoryApiFp(this.configuration).inventoryControllerGetItemV1(requestParameters.id, options).then((request) => request(this.axios, this.basePath));
    }

    /**
     * 
     * @param {InventoryApiInventoryControllerGetItemsV1Request} requestParameters Request parameters.
     * @param {*} [options] Override http request option.
     * @throws {RequiredError}
     */
    public inventoryControllerGetItemsV1(requestParameters: InventoryApiInventoryControllerGetItemsV1Request = {}, options?: RawAxiosRequestConfig) {
        return InventoryApiFp(this.configuration).inventoryControllerGetItemsV1(requestParameters.sort, requestParameters.skip, requestParameters.limit, requestParameters.query, options).then((request) => request(this.axios, this.basePath));
    }

    /**
     * 
     * @param {InventoryApiInventoryControllerUploadItemImageV1Request} requestParameters Request parameters.
     * @param {*} [options] Override http request option.
     * @throws {RequiredError}
     */
    public inventoryControllerUploadItemImageV1(requestParameters: InventoryApiInventoryControllerUploadItemImageV1Request, options?: RawAxiosRequestConfig) {
        return InventoryApiFp(this.configuration).inventoryControllerUploadItemImageV1(requestParameters.id, requestParameters.image, options).then((request) => request(this.axios, this.basePath));
    }
}